requirements.txt
requests>=2.31.0
gql>=3.4.1
requests-toolbelt>=1.0.0
pandas>=2.0.0
numpy>=1.24.0
scipy>=1.11.0
matplotlib>=3.7.0
seaborn>=0.12.0
python-dotenv>=1.0.0

